Split header account state
